requirements-client.txt
aiogram==3.10.0
asyncpg==0.29.0
httpx==0.27.2
python-dotenv==1.0.1
redis==5.0.7
